+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/AnnKitTest.groovy
@@ -52,6 +52,44 @@ class AnnKitTest extends AbstractKitTest {
     }
 
 
+    def "Check enclosing class annotation detection for inner class"() {
+        setup:
+        build """
+            plugins {
+                id 'java'
+                id 'ru.vyarus.animalsniffer'
+            }
+
+            animalsniffer {
+                ignoreFailures = true
+            }
+
+            repositories { mavenCentral()}
+            dependencies {
+                signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
+                compile "org.codehaus.mojo:animal-sniffer-annotations:1.14"
+            }
+
+        """
+        fileFromClasspath('src/main/java/ann/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/ann4inner/Sample.java')
+        fileFromClasspath('src/main/java/ann/Sample2.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/ann4inner/Sample2.java')
+//        debug()
+
+        when: "run task"
+        BuildResult result = run('check')
+
+        then: "task successful"
+        result.task(':check').outcome == TaskOutcome.SUCCESS
+
+        then: "found no violations"
+        !result.output.contains("AnimalSniffer violations were found")
+
+        then: "report correct"
+        File file = file('/build/reports/animalsniffer/main.text')
+        !file.exists()
+    }
+
+
     def "Check custom annotation detection"() {
         setup:
         build """